+++ microsoft-edge/devtools-guide-chromium/whats-new/whats-new.md
@@ -5,17 +5,20 @@ author: MSEdgeTeam
 ms.author: msedgedevrel
 ms.topic: conceptual
 ms.prod: microsoft-edge
-ms.date: 03/02/2023
+ms.date: 08/18/2023
 ---
 # What's New in Microsoft Edge DevTools
 
 [!INCLUDE [Microsoft Edge team note for top of What's New](includes/edge-whats-new-note.md)]
+
+What's New articles correspond to historical releases of Microsoft Edge, and are not updated over time.  References to "new features" and "experiments" might become outdated as new versions of Microsoft Edge are released.
 
 <!-- maintenance notes:
 * add the new page to toc.yml
 * move eleventh oldest link into whats-new-archive.md
 -->
 
+* [What's New in DevTools (Microsoft Edge 115)](2023/07/devtools-115.md)
 * [What's New in DevTools (Microsoft Edge 114)](2023/06/devtools-114.md)
 * [What's New in DevTools (Microsoft Edge 113)](2023/05/devtools-113.md)
 * [What's New in DevTools (Microsoft Edge 112)](2023/04/devtools-112.md)
@@ -25,5 +28,4 @@ ms.date: 03/02/2023
 * [What's New in DevTools (Microsoft Edge 108)](2022/12/devtools-108.md)
 * [What's New in DevTools (Microsoft Edge 107)](2022/10/devtools-107.md)
 * [What's New in DevTools (Microsoft Edge 106)](2022/09/devtools-106.md)
-* [What's New in DevTools (Microsoft Edge 105)](2022/09/devtools-105.md)
 * [What's New in DevTools archive](./whats-new-archive.md)